apps/worker/requirements.txt
# ── Web framework ──
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── LLM provider abstraction ──
litellm>=1.30.0

# ── Data processing ──
pandas>=2.1.0

# ── Configuration ──
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ── Database (mirrors Prisma schema) ──
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# ── Azure Storage SDK ──
azure-storage-queue>=12.8.0
azure-storage-blob>=12.19.0

# ── Templating (prompts) ──
jinja2>=3.1.0
pyyaml>=6.0

# ── Testing ──
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
